Added TieGroupFilter, to reduce permutation choices
This postulates that any choices that would tie in a head on head
election must also tie in the final results. Thus, a "group leader" is
selected to be run on the backup algorithm, then the entire group is
replaced for that leader in the results.

I don't know if the algorithm really supports this or not...